Various fixes to compile with VS2012
Fixes mostly required for all the new code which came in with the
merge of the hcloud stuff:

* Silence and fix a bunch of warnings regarding implicit integer and
  floating point conversions.

* Work around lack of variadic templates for displaz C++ bindings

* Work around broken autogenerated move constructor & assignment

+++ src/dvox.cpp
@@ -58,7 +58,7 @@ int main(int argc, char* argv[])
 {
     Imath::V3d boundMin = Imath::V3d(0);
     double rootNodeWidth = 1000;
-    float pointRadius = 0.2;
+    float pointRadius = 0.2f;
     int brickRes = 8;
     double leafNodeWidth = 2.5;
 
@@ -147,7 +147,7 @@ int main(int argc, char* argv[])
                                   (size_t)(dbCacheSize*1024*1024),
                                   logger);
 
-            int leafDepth = floor(log(rootNodeWidth/leafNodeWidth)/log(2) + 0.5);
+            int leafDepth = (int)floor(log(rootNodeWidth/leafNodeWidth)/log(2) + 0.5);
             logger.info("Leaf node width = %.3f", rootNodeWidth / (1 << leafDepth));
             std::ofstream outputFile(outputPath);
             voxelizePointCloud(outputFile, pointDb, pointRadius,
