+++ src/view3d.cpp
@@ -545,7 +545,7 @@ DrawCount View3D::drawPoints(const TransformState& transState,
             continue;
         V3f relCursor = m_cursorPos - geom.offset();
         prog.setUniformValue("cursorPos", relCursor.x, relCursor.y, relCursor.z);
-        prog.setUniformValue("fileNumber", (GLint)(selection[i].row() + 1));
+        prog.setUniformValue("fileNumber", (GLint)(selection[(int)i].row() + 1));
         prog.setUniformValue("pointPixelScale", (GLfloat)(0.5*width()*m_camera.projectionMatrix()[0][0]));
         totDrawCount += geom.drawPoints(prog, transState, quality, incrementalDraw);
     }
